Change version field from int to byte for ClusterConfig and ReplicationHistory
Revert version type from int (4 bytes) to byte (1 byte) for both
ClusterConfig and ReplicationHistory serialization. The binary layout
is now: version byte (1 byte) + payload, minimizing wire overhead.

+++ libs/cluster/Server/ClusterConfig.cs
@@ -44,7 +44,7 @@ namespace Garnet.cluster
         /// Version of the cluster config serialization format.
         /// Increment when the binary layout of <see cref="ToByteArray"/>/<see cref="FromByteArray"/> changes.
         /// </summary>
-        public const int ClusterConfigVersion = 1;
+        public const byte ClusterConfigVersion = 1;
 
         /// <summary>
         ///
